Conflicts:
	nds-client/src/main/webapp/WEB-INF/jsp/cotaGarantia/index.jsp

+++ nds-model/src/main/java/br/com/abril/nds/dto/StatusDividaDTO.java
@@ -156,7 +156,7 @@ public class StatusDividaDTO implements Serializable {
 		return diasAtraso;
 	}
 
-	public void setDiasAtraso(Integer diasAtraso) {
+	public void setDiasAtraso(BigInteger diasAtraso) {
 		this.diasAtraso = diasAtraso.longValue();
 	}
